No need to create orgin file if existed
If crc error, the orgin file is saved for analysis. Then if
we re-run rwcheck, it will modify the old orgin file, which
is not what we want.

+++ rwcheck.c
@@ -347,6 +347,9 @@ int init_orgin_file(struct rwtask *task)
 	if (task->set_org)
 		return 0;
 
+	if (is_existed(task->org))
+		return 0;
+
 	/* create orgin file */
 	return copy_file(RANDOM_DATA, task->org, DEFAULT_ORG_SIZE, 4 * KB);
 }
